users/pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = ""
authors = ["Aleksei Kurkovskii <alkurkovsk@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlalchemy = "^2.0.18"
asyncpg = "^0.27.0"
alembic = "^1.11.1"
gunicorn = "^20.1.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
black = "^24.4.2"
reorder-python-imports = "^3.13.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
